backend/requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==4.2.7
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-filter==23.3
djangorestframework==3.14.0
iniconfig==2.0.0
Markdown==3.5.1
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
pytest==7.4.3
pytest-django==4.7.0
pytz==2023.3.post1
sorl-thumbnail==12.10.0
sqlparse==0.4.4
tzdata==2023.3
